#22 move: oauth, auth 타입 분리
