+++ src/types/auth.ts
@@ -21,14 +21,3 @@ export interface LoginParams {
 export interface SignupParams extends LoginParams {
   nickname: string;
 }
-
-export type OAuthProvider = 'google' | 'kakao';
-
-export interface OauthSigninParams {
-  provider: OAuthProvider;
-  code: string | null;
-}
-
-export interface OauthSignupParams extends OauthSigninParams {
-  nickname: string;
-}
